A nonparametric estimator for lagged identification rate

A nonparametric estimator for lagged identification rate

Usage

lir_nonparametric_estimation(X, n, tp)

Arguments

X

A list or matrix containing the identities of individuals identified during each sampling period

n

A vector or a positive integer, representing the number of individuals identified in each sampling period. It indicates the same number of individuals identified in all sampling periods if a positive integer.

tp

A set of observation time

Value

A list with the following elements:

tau

a set of time lags

R_tau

A nonparametric estimator of \(\hat{R}(\tau)\) was given by Whitehead (2007)

R_m

\(\sum_{i,j} \{m_{t_i,t_j}|\tau_{ij}=\tau\}\), with \(\tau_{ij}=|t_i-t_j|\)

R_n

\(\sum_{i,j} \{n_i*n_j|\tau_{ij}=\tau\}\), with \(\tau_{ij}=|t_i-t_j|\)

mij

number of individuals identified at both time \(t_i\) and \(t_j\)

nij

\(n_i\)*\(n_j\)

tauij

\(\tau_{ij}=|t_i-t_j|\)

Details

The lagged identification rate \(R(\tau)\) is the probability that an individual in the study area is identified now and is reidentified after a time lag of \(\tau\). A nonparametric estimator of \(R(\tau)\) was given by Whitehead (2007): $$\hat{R}(\tau)=\frac{\sum_{i,j} \{m_{t_i,t_j}|\tau_{ij}=\tau\}}{\sum_{i,j}\{n_i*n_j|\tau_{ij}=\tau\}},$$ where where \(m_{t_i,t_j}\) is the number of individuals identified at both time \(t_i\) and \(t_j\), and \(n_{t_i}\) represents the number of individuals identified at time \(t_i\).

Examples

# Example
tp <- c(1:5, 51:55, 101:105, 501:505, 601:605)
N <- 100; n <- 10
X <- list()
for (i in tp){
X[[i]] <- sample(1:N, n)
}

res <- lir_nonparametric_estimation(X, n, tp)
res$R_tau; res$tau
#>  [1] 0.011500000 0.010000000 0.012000000 0.010000000 0.014000000 0.011250000
#>  [7] 0.011666667 0.007500000 0.000000000 0.011000000 0.011250000 0.018333333
#> [13] 0.010000000 0.005000000 0.010000000 0.007500000 0.006666667 0.007500000
#> [19] 0.020000000 0.008000000 0.005000000 0.000000000 0.000000000 0.000000000
#> [25] 0.008750000 0.010000000 0.007500000 0.005000000 0.006666667 0.008333333
#> [31] 0.011666667 0.006666667 0.007500000 0.012500000 0.010000000 0.010000000
#> [37] 0.020000000 0.005000000 0.010000000 0.010000000 0.012000000 0.007500000
#> [43] 0.013333333 0.005000000 0.010000000 0.012000000 0.010000000 0.013333333
#> [49] 0.000000000 0.020000000 0.020000000 0.007500000 0.016666667 0.003333333
#> [55] 0.010000000 0.005000000 0.000000000 0.010000000 0.010000000 0.005000000
#> [61] 0.013333333 0.015000000 0.010000000 0.005000000 0.003333333 0.015000000
#> [67] 0.010000000
#>  [1]   1   2   3   4  50  51  52  53  54 100 101 102 103 104 500 501 502 503 504
#> [20] 600 601 602 603 604  49  99 499 599  48  98 498 598  47  97 497 597  46  96
#> [39] 496 596 450 451 452 453 454 550 551 552 553 554 449 549 448 548 447 547 446
#> [58] 546 400 401 402 403 404 399 398 397 396